requirements.txt
python-telegram-bot==21.3
apscheduler==3.10.4
requests==2.32.3
beautifulsoup4==4.12.3
